chore: refactor getSubsidyToApplyForCourse and constants location
