+++ src/components/course/data/constants.js
@@ -10,10 +10,10 @@ export const SUBSIDY_DISCOUNT_TYPE_MAP = {
   ABSOLUTE: 'absolute',
 };
 
-export const LICENSE_SUBSIDY_TYPE = 'license';
-export const COUPON_CODE_SUBSIDY_TYPE = 'couponCode';
-export const ENTERPRISE_OFFER_SUBSIDY_TYPE = 'enterpriseOffer';
-export const LEARNER_CREDIT_SUBSIDY_TYPE = 'learnerCredit';
+// export const LICENSE_SUBSIDY_TYPE = 'license';
+// export const COUPON_CODE_SUBSIDY_TYPE = 'couponCode';
+// export const ENTERPRISE_OFFER_SUBSIDY_TYPE = 'enterpriseOffer';
+// export const LEARNER_CREDIT_SUBSIDY_TYPE = 'learnerCredit';
 
 export const PROMISE_FULFILLED = 'fulfilled';
