+++ src/components/course/routes/tests/ExternalCourseEnrollment.test.jsx
@@ -7,8 +7,6 @@ import { renderWithRouterProvider } from '../../../../utils/tests';
 
 import {
   DISABLED_ENROLL_REASON_TYPES,
-  LEARNER_CREDIT_SUBSIDY_TYPE,
-  LICENSE_SUBSIDY_TYPE,
 } from '../../data/constants';
 import ExternalCourseEnrollment from '../ExternalCourseEnrollment';
 import { CourseContext } from '../../CourseContextProvider';
@@ -16,6 +14,8 @@ import {
   useCourseRedemptionEligibility,
   useEnterpriseCourseEnrollments,
   useEnterpriseCustomer,
+  LEARNER_CREDIT_SUBSIDY_TYPE,
+  LICENSE_SUBSIDY_TYPE,
 } from '../../../app/data';
 import { enterpriseCustomerFactory } from '../../../app/data/services/data/__factories__';
 import {
