+++ src/components/dashboard/main-content/course-enrollments/data/hooks.js
@@ -17,7 +17,6 @@ import {
   createEnrollWithLicenseUrl,
   findCouponCodeForCourse,
   findHighestLevelSeatSku,
-  getSubsidyToApplyForCourse,
 } from '../../../../course/data/utils';
 import { getHasUnacknowledgedAssignments } from '../../../data/utils';
 import { ASSIGNMENT_TYPES } from '../../../../enterprise-user-subsidy/enterprise-offers/data/constants';
@@ -33,6 +32,7 @@ import {
   useCanUpgradeWithLearnerCredit,
   useSubscriptions,
   useCouponCodes,
+  getSubsidyToApplyForCourse,
 } from '../../../../app/data';
 import {
   sortedEnrollmentsByEnrollmentDate,
